add soft delete

+++ src/BackupDatabase.php
@@ -29,11 +29,12 @@ class BackupDatabase
                 $username = $connection['db_username'] ?? config("database.connections.{$connectionDatabase}.username");
                 $password = $connection['db_password'] ?? config("database.connections.{$connectionDatabase}.password");
                 $day_for_delete = $connection['day_for_delete'] ?? null;
+                $soft_delete = $connection['soft_delete'] ?? false;
                 $daily = $connection['daily'];
                 $destinationpath = $connection['destinationpath'];
                 $name = $dbname  .  ($daily ? "_" . Carbon::now()->format($connection['datetimeFormat']) : "") . ".bak";
 
-                $resultPrevius = $this->checkPreviousBackups($destinationpath, $dbname, $day_for_delete);
+                $resultPrevius = $this->checkPreviousBackups($destinationpath, $dbname, $day_for_delete, $soft_delete);
 
                 $resultCreateFolder = $this->createFolder($destinationpath);
                 if ($resultCreateFolder['status'] == false) {
@@ -95,21 +96,20 @@ class BackupDatabase
         return false;
     }
 
-    private function checkPreviousBackups($destinationpath, $dbname, $day_for_delete)
+    private function checkPreviousBackups($destinationpath, $dbname, $day_for_delete, $soft_delete)
     {
-
         if ($day_for_delete === null) {
             return [];
         }
         $result = null;
         foreach (glob($destinationpath . $dbname . "*.bak") as $file) {
-            $result[] = $this->deleteAfter($day_for_delete, $file);
+            $result[] = $this->deleteAfter($day_for_delete, $file, $soft_delete);
         }
 
         return array_filter(is_array($result) ? $result : []);
     }
 
-    function deleteAfter($day_for_delete, $filename)
+    function deleteAfter($day_for_delete, $filename, $soft_delete)
     {
         if (!file_exists($filename)) {
             return ['status' => false, 'message' => "File {$filename} not found."];
@@ -118,8 +118,21 @@ class BackupDatabase
         $date_now_sub_for_delate = Carbon::now()->subDays($day_for_delete);
 
         if ($date_now_sub_for_delate > $date_file) {
+            if ($soft_delete) {
+                $fileinfo = pathinfo($filename);
 
-            unlink($filename);
+                $trash = $fileinfo['dirname']  . '\\trash\\';
+                $resultCreateFolder = $this->createFolder($trash);
+
+                if ($resultCreateFolder['status'] == false) {
+                    return ['status' => false, 'message' => "Folder {$trash}  not writable."];
+                }
+                $filenameTrash = $trash . basename($filename);
+                rename($filename, $filenameTrash);
+            } else {
+                unlink($filename);
+            }
+
             return ['status' => true, 'message' => "File {$filename} deleted."];
         }
     }
